+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - Support objects as parameters to functions via GraphQL input types [#359](https://github.com/hypermodeAI/runtime/pull/359)
 - Fix GraphQL schema generation for Go functions [#360](https://github.com/hypermodeAI/runtime/pull/360)
 - Fix import registration issue [#364](https://github.com/hypermodeAI/runtime/pull/364)
+- Fix conversion of empty arrays and slices [#365](https://github.com/hypermodeAI/runtime/pull/365)
 
 ## 2024-08-27 - Version 0.11.2